# docs/openapi.yaml
openapi: "3.0.3"
info:
  title: XLAIR API
  description: ""
  license:
    name: ""
  version: "1.0.0"
tags:
  - name: app
    description: XLAIR の筐体
  - name: station
    description: カード情報登録機
  - name: web
    description: xlair.dev
  - name: admin
    description: 管理用/デバッグ用
  - name: common
    description: 共通の API
servers:
  - url: https://api.xlair.dev/v1
paths:
  /users:
    post:
      tags:
        - station
      summary: ユーザー登録
      description: カードIDで筐体に新規ログインした時に呼ぶ
      security:
        - deviceAuth: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/userRegister"
      responses:
        "201":
          description: User created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/userData"
        "400":
          description: Bad request - Invalid input data
        "401":
          description: Unauthorized - Invalid device access token
        "409":
          description: Conflict - User with this card ID already exists
        "500":
          description: Internal server error
    get:
      tags:
        - app
        - station
      summary: カードIDからユーザー情報を取得
      description: カードIDで筐体に通常ログインした時に呼ぶ
      security:
        - deviceAuth: []
      parameters:
        - name: card
          in: query
          description: カードID
          required: true
          schema:
            type: string
      responses:
        "200":
          description: success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/userData"
        "400":
          description: Bad request - Missing or invalid card ID
        "401":
          description: Unauthorized - Invalid device access token
        "404":
          description: Not found - User with specified card ID not found
        "500":
          description: Internal server error
  /users/{userId}:
    patch:
      tags:
        - station
      summary: ユーザー情報の更新
      description: ユーザーの表示名と公開設定を更新する
      security:
        - deviceAuth: []
      parameters:
        - name: userId
          in: path
          description: ユーザーのID
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/userUpdate"
      responses:
        "200":
          description: success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/userData"
        "401":
          description: Unauthorized - Invalid access token
        "403":
          description: Forbidden - The authenticated principal cannot update this user
        "404":
          description: Not found - User not found
        "500":
          description: Internal server error
  /admin/db/synchronize:
    post:
      tags:
        - admin
      summary: DBの派生値を同期
      description: 現在のドメインルールに基づいて、DBに保存された派生値を同期する。値が変わるユーザーのみ更新する。
      security:
        - userAuth: []
      responses:
        "200":
          description: Recalculation completed
          content:
            application/json:
              schema:
                type: object
                properties:
                  updatedUsers:
                    type: integer
                    format: int64
                  updatedRatings:
                    type: integer
                    format: int64
                required:
                  - updatedUsers
                  - updatedRatings
        "401":
          description: Unauthorized - Invalid user access token
        "403":
          description: Forbidden - The authenticated user is not an administrator
        "500":
          description: Internal server error
  /admin/musics:
    post:
      tags:
        - admin
      summary: 楽曲と譜面を登録
      security:
        - userAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              allOf:
                - $ref: "#/components/schemas/musicData"
                - type: object
                  properties:
                    sheets:
                      type: array
                      minItems: 3
                      maxItems: 3
                      items:
                        $ref: "#/components/schemas/sheetData"
                  required:
                    - sheets
      responses:
        "201":
          description: 登録された楽曲と譜面
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/musicWithSheets"
        "400":
          description: Bad request - Invalid music or sheet data
        "401":
          description: Unauthorized - Invalid access token
        "403":
          description: Forbidden - The authenticated user is not an administrator
        "500":
          description: Internal server error
    get:
      tags:
        - admin
      summary: 楽曲と譜面の一覧を取得
      description: 楽曲を登録日時とIDの昇順で取得する。cursorを指定すると続きから取得する。
      security:
        - userAuth: []
      parameters:
        - name: cursor
          in: query
          description: 前ページのレスポンスに含まれる次のページ用カーソル
          required: false
          schema:
            type: string
        - name: limit
          in: query
          description: 取得件数 (1〜100、既定値50)
          required: false
          schema:
            type: integer
            format: int64
            minimum: 1
            maximum: 100
            default: 50
      responses:
        "200":
          description: 楽曲と譜面のページ
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/adminMusicListResponse"
        "400":
          description: Bad request - Invalid cursor or limit
        "401":
          description: Unauthorized - Invalid user access token
        "403":
          description: Forbidden - The authenticated user is not an administrator
        "500":
          description: Internal server error
  /admin/musics/{musicId}:
    get:
      tags:
        - admin
      summary: 楽曲と譜面の詳細を取得
      security:
        - userAuth: []
      parameters:
        - name: musicId
          in: path
          description: 楽曲のID
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: 楽曲と譜面の詳細
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/musicWithSheets"
        "401":
          description: Unauthorized - Invalid access token
        "403":
          description: Forbidden - The authenticated user is not an administrator
        "404":
          description: Not found - Music not found
        "500":
          description: Internal server error
    post:
      tags:
        - admin
      summary: 楽曲と譜面を更新
      security:
        - userAuth: []
      parameters:
        - name: musicId
          in: path
          description: 楽曲のID
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              allOf:
                - $ref: "#/components/schemas/musicData"
                - type: object
                  properties:
                    sheets:
                      type: array
                      minItems: 3
                      maxItems: 3
                      items:
                        $ref: "#/components/schemas/sheet"
                  required:
                    - sheets
      responses:
        "200":
          description: 更新された楽曲と譜面
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/musicWithSheets"
        "400":
          description: Bad request - Invalid music or sheet data
        "401":
          description: Unauthorized - Invalid access token
        "403":
          description: Forbidden - The authenticated user is not an administrator
        "404":
          description: Not found - Music not found
        "500":
          description: Internal server error
  /admin/jackets/upload-url:
    post:
      tags:
        - admin
      summary: 楽曲ジャケットのアップロード URL を発行
      security:
        - userAuth: []
      parameters:
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/jacketUploadRequest"
      responses:
        "200":
          description: R2へのアップロード URL
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/jacketUploadResponse"
        "400":
          description: Bad request - Invalid music ID or content type
        "401":
          description: Unauthorized - Invalid access token
        "403":
          description: Forbidden - The authenticated user is not an administrator
        "500":
          description: Internal server error
  /users/{userId}/credits/increment:
    post:
      tags:
        - app
      summary: ユーザーのクレジット数を更新
      description: ユーザーのクレジット数を1増やす。ゲーム開始時に呼ぶ
      security:
        - deviceAuth: []
      parameters:
        - name: userId
          in: path
          description: ユーザーのID
          required: true
          schema:
            type: string
      responses:
        "200":
          description: success
          content:
            application/json:
              schema:
                type: object
                properties:
                  credits:
                    type: integer
                    description: 更新後のクレジット数
        "401":
          description: Unauthorized - Invalid device access token
        "404":
          description: Not found - User not found
        "500":
          description: Internal server error
  /users/{userId}/records:
    get:
      tags:
        - app
      summary: ユーザーのプレイデータを取得
      description: ゲーム開始時にユーザーのプレイデータを一括取得する
      security:
        - deviceAuth: []
      parameters:
        - name: userId
          in: path
          description: ユーザーのID
          required: true
          schema:
            type: string
      responses:
        "200":
          description: success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/record"
        "401":
          description: Unauthorized - Invalid device access token
        "404":
          description: Not found - User not found
        "500":
          description: Internal server error
    post:
      tags:
        - app
      summary: ユーザーのプレイデータを送信
      description: スコアの更新の有無に関わらず、ゲーム終了時にユーザーのプレイデータを送信する
      security:
        - deviceAuth: []
      parameters:
        - name: userId
          in: path
          description: ユーザーのID
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: "#/components/schemas/recordRequest"
      responses:
        "201":
          description: Records created/updated successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/record"
        "400":
          description: Bad request - Invalid input data
        "401":
          description: Unauthorized - Invalid device access token
        "404":
          description: Not found - User or sheet not found
        "500":
          description: Internal server error
  /users/{userId}/options:
    get:
      tags:
        - app
      summary: ユーザーのプレイオプションを取得
      description: ゲーム開始時にユーザーのプレイオプションを取得する
      security:
        - deviceAuth: []
      parameters:
        - name: userId
          in: path
          description: ユーザーのID
          required: true
          schema:
            type: string
      responses:
        "200":
          description: success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/userPlayOption"
        "401":
          description: Unauthorized - Invalid device access token
        "404":
          description: Not found - User not found
        "500":
          description: Internal server error
    post:
      tags:
        - app
      summary: ユーザーのプレイオプションを更新
      description: ゲーム終了時にユーザーのプレイオプションを保存する
      security:
        - deviceAuth: []
      parameters:
        - name: userId
          in: path
          description: ユーザーのID
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/userPlayOption"
      responses:
        "200":
          description: success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/userPlayOption"
        "401":
          description: Unauthorized - Invalid device access token
        "404":
          description: Not found - User not found
        "500":
          description: Internal server error
  /sync:
    get:
      tags:
        - app
      summary: 全曲・全譜面のメタデータを取得
      description: 同期用。筐体が起動した最初の一回しか実行しない
      security:
        - deviceAuth: []
      responses:
        "200":
          description: success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/musicWithSheets"
        "401":
          description: Unauthorized - Invalid device access token
        "500":
          description: Internal server error
  /rankings/sheets/{sheetId}:
    get:
      tags:
        - web
      summary: 指定した譜面のハイスコアランキングを取得
      description: 公開ユーザーを対象に、指定した譜面のハイスコア上位20件を取得する
      parameters:
        - name: sheetId
          in: path
          description: 譜面のID
          required: true
          schema:
            type: string
      responses:
        "200":
          description: success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/sheetScoreRankingResponse"
        "404":
          description: Not found - Sheet not found
        "500":
          description: Internal server error
  /rankings/total-score:
    get:
      tags:
        - web
      summary: 全譜面のハイスコア合計ランキングを取得
      description: 公開ユーザーを対象に、保有する全譜面のハイスコア合計上位20件を取得する
      responses:
        "200":
          description: success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/totalScoreRankingResponse"
        "500":
          description: Internal server error
  /rankings/rating:
    get:
      tags:
        - web
      summary: レーティングのランキングを取得
      description: 公開ユーザーを対象に、レーティングの上位20件を取得する
      responses:
        "200":
          description: success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ratingRankingResponse"
        "500":
          description: Internal server error
  /rankings/xp:
    get:
      tags:
        - web
      summary: 経験値のランキングを取得
      description: 公開ユーザーを対象に、XP の上位20件を取得する
      responses:
        "200":
          description: success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/xpRankingResponse"
        "500":
          description: Internal server error
  /statistics/summary:
    get:
      tags:
        - web
      summary: システム全体の統計情報を取得
      description: 全ユーザーとレコードの集計値を参照する
      responses:
        "200":
          description: success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/globalStatistics"
        "500":
          description: Internal server error
  /health:
    get:
      tags:
        - common
      summary: ヘルスチェック
      description: API サーバーの稼働状況を確認する
      responses:
        "200":
          description: API サーバーが正常に稼働中
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/healthCheck"
        "500":
          description: Internal server error
components:
  securitySchemes:
    userAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Auth0 によるユーザー認証トークン。
        現時点では管理者・運用者用 dashboard で利用する。
        一般ユーザーは Auth0 アカウントを作成せず、筐体では card ID で識別する。
    deviceAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Auth0 Machine-to-Machine Application による筐体認証トークン。
        XLAIR が認証主体を筐体として分類するための大分類を含む。
  schemas:
    userRegister:
      type: object
      properties:
        card:
          type: string
          description: カードID
        displayName:
          type: string
          description: ユーザーの表示名
        isPublic:
          type: boolean
          description: ユーザー情報を xlair.dev 上に公開するかどうかのフラグ
      required:
        - card
        - displayName
        - isPublic
    userUpdate:
      type: object
      properties:
        displayName:
          type: string
          description: ユーザーの表示名
        isPublic:
          type: boolean
          description: ユーザー情報を xlair.dev 上に公開するかどうかのフラグ
      required:
        - displayName
        - isPublic
    userData:
      type: object
      properties:
        id:
          type: string
          description: ユーザーのID
        card:
          type: string
          description: Felica の IDm
        displayName:
          type: string
          description: ユーザーの表示名
        rating:
          type: integer
          description: ユーザーのレーティング
        xp:
          type: integer
          description: ユーザーの経験値
        credits:
          type: integer
          description: ユーザーの総クレジット数
        isPublic:
          type: boolean
          description: ユーザーの情報を xlair.dev 上に公開するかどうかのフラグ
        isAdmin:
          type: boolean
          description: ユーザーが管理者かどうかを表すフラグ
        createdAt:
          type: string
          format: date-time
          description: ユーザーの登録日時
      required:
        - id
        - card
        - displayName
        - rating
        - xp
        - credits
        - createdAt
    userPlayOption:
      type: object
      properties:
        noteSpeed:
          type: number
          format: float
          description: ノーツの流れる速度
        judgmentOffset:
          type: integer
          description: 判定タイミングの調整値 (ミリ秒)
      required:
        - noteSpeed
        - judgmentOffset
    recordRequest:
      type: object
      properties:
        userId:
          type: string
          description: ユーザーのID
        sheetId:
          type: string
          description: 譜面のID
        score:
          type: integer
          description: ハイスコア
        clearType:
          type: string
          enum:
            - failed
            - clear
            - fullcombo
            - perfect
          description: クリアタイプ
      required:
        - userId
        - sheetId
        - score
        - clearType
    record:
      type: object
      properties:
        id:
          type: string
          description: レコードのID
        userId:
          type: string
          description: ユーザーのID
        sheetId:
          type: string
          description: 譜面のID
        score:
          type: integer
          description: ハイスコア
        clearType:
          type: string
          enum:
            - failed
            - clear
            - fullcombo
            - perfect
          description: クリアタイプ
        playCount:
          type: integer
          description: プレイ回数
        updatedAt:
          type: string
          format: date-time
          description: レコードの更新日時
      required:
        - id
        - userId
        - sheetId
        - score
        - clearType
        - playCount
        - updatedAt
    jacketUploadRequest:
      type: object
      properties:
        musicId:
          type: string
          format: uuid
          description: 編集対象の楽曲 ID。新規登録時は省略する
        contentType:
          type: string
          enum:
            - image/jpeg
            - image/png
            - image/webp
          description: アップロードする画像の MIME タイプ
      required:
        - contentType
    jacketUploadResponse:
      type: object
      properties:
        uploadUrl:
          type: string
          format: uri
          description: R2 への期限付き PUT URL
        jacketUrl:
          type: string
          format: uri
          description: 楽曲に設定する公開 URL
      required:
        - uploadUrl
        - jacketUrl
    musicData:
      type: object
      properties:
        title:
          type: string
          description: タイトル
        artist:
          type: string
          description: アーティスト名
        bpm:
          type: number
          format: float
          description: BPM
        genre:
          type: string
          enum:
            - ORIGINAL
          description: ジャンル
        jacket:
          type: string
          description: ジャケット画像。プロジェクトのルートからの相対パス、または xlair.dev/public/ からの絶対パス
        registrationDate:
          type: string
          format: date-time
          description: 運用上の楽曲登録日
        isTest:
          type: boolean
          description: テスト楽曲かどうか
      required:
        - title
        - artist
        - bpm
        - genre
        - jacket
        - registrationDate
        - isTest
    music:
      allOf:
        - $ref: "#/components/schemas/musicData"
        - type: object
          properties:
            id:
              type: string
              description: 楽曲のID
          required:
            - id
    sheetData:
      type: object
      properties:
        difficulty:
          type: string
          enum:
            - easy
            - normal
            - hard
          description: 難易度の分類
        level:
          type: number
          description: レベル
        notesDesigner:
          type: string
          description: 譜面のノーツデザイナー
      required:
        - difficulty
        - level
        - notesDesigner
    sheetWithMusicId:
      allOf:
        - $ref: "#/components/schemas/sheetData"
        - type: object
          properties:
            id:
              type: string
              description: 譜面のID
            musicId:
              type: string
              description: 楽曲のID
          required:
            - id
            - musicId
    sheet:
      allOf:
        - $ref: "#/components/schemas/sheetData"
        - type: object
          properties:
            id:
              type: string
              format: uuid
              description: 譜面のID
          required:
            - id
    musicWithSheets:
      type: object
      properties:
        music:
          $ref: "#/components/schemas/music"
        sheets:
          type: array
          items:
            $ref: "#/components/schemas/sheetWithMusicId"
      required:
        - music
        - sheets
    adminMusicListResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/musicWithSheets"
        nextCursor:
          type: string
          nullable: true
      required:
        - items
        - nextCursor
    healthCheck:
      type: object
      properties:
        status:
          type: string
          example: "ok"
          description: API サーバーの稼働状況
        timestamp:
          type: string
          format: date-time
          description: レスポンス生成時刻
      required:
        - status
        - timestamp
    globalStatistics:
      type: object
      properties:
        totalCredits:
          type: integer
          description: 全ユーザーのプレイ数 (クレジット) の合計
        totalUsers:
          type: integer
          description: 登録済みユーザー数
        totalScore:
          type: integer
          description: 全レコードのハイスコア合計
      required:
        - totalCredits
        - totalUsers
        - totalScore
    sheetScoreRankingEntry:
      type: object
      properties:
        rank:
          type: integer
          description: ランキング順位 (1 始まり)
        userId:
          type: string
          description: ユーザーのID
        displayName:
          type: string
          description: ユーザーの表示名
        score:
          type: integer
          description: 譜面におけるハイスコア
      required:
        - rank
        - userId
        - displayName
        - score
    sheetScoreRankingResponse:
      type: object
      properties:
        sheetId:
          type: string
          description: 取得対象の譜面ID
        entries:
          type: array
          description: ランキング上位20件
          items:
            $ref: "#/components/schemas/sheetScoreRankingEntry"
      required:
        - sheetId
        - entries
    totalScoreRankingEntry:
      type: object
      properties:
        rank:
          type: integer
          description: ランキング順位 (1 始まり)
        userId:
          type: string
          description: ユーザーのID
        displayName:
          type: string
          description: ユーザーの表示名
        totalScore:
          type: integer
          description: 全譜面のハイスコア合計
      required:
        - rank
        - userId
        - displayName
        - totalScore
    totalScoreRankingResponse:
      type: object
      properties:
        entries:
          type: array
          description: ランキング上位20件
          items:
            $ref: "#/components/schemas/totalScoreRankingEntry"
      required:
        - entries
    ratingRankingEntry:
      type: object
      properties:
        rank:
          type: integer
          description: ランキング順位 (1 始まり)
        userId:
          type: string
          description: ユーザーのID
        displayName:
          type: string
          description: ユーザーの表示名
        rating:
          type: integer
          description: レーティング値
      required:
        - rank
        - userId
        - displayName
        - rating
    ratingRankingResponse:
      type: object
      properties:
        entries:
          type: array
          description: ランキング上位20件
          items:
            $ref: "#/components/schemas/ratingRankingEntry"
      required:
        - entries
    xpRankingEntry:
      type: object
      properties:
        rank:
          type: integer
          description: ランキング順位 (1 始まり)
        userId:
          type: string
          description: ユーザーのID
        displayName:
          type: string
          description: ユーザーの表示名
        xp:
          type: integer
          description: 獲得済みの経験値
      required:
        - rank
        - userId
        - displayName
        - xp
    xpRankingResponse:
      type: object
      properties:
        entries:
          type: array
          description: ランキング上位20件
          items:
            $ref: "#/components/schemas/xpRankingEntry"
      required:
        - entries